refactor(deepseek-v4): move fused SwiGLU kernel to tokenspeed-kernel
Move _fused_swiglu_fp8_ue8m0_kernel from runtime models/ to
tokenspeed_kernel/ops/activation/triton.py where kernel code belongs
per AGENTS.md.  Runtime import updated to tokenspeed_kernel path.
Delete the now-empty deepseek_v4_fused_swiglu.py.

Also make need_scores automatic: _select_experts now checks
self.experts.topk_output_format instead of requiring callers to
pass need_scores manually.  All current backends use BYPASSED
format, and mega_moe has no topk_output_format attr, so scores
are skipped in both mega_moe and forward_normal paths.

GSM8K 5-shot 50-sample: 0.98 (unchanged).
